Unnecessary cast
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/vfs/trunk@1030177 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/commons/vfs/provider/tar/TarInputStream.java
@@ -437,7 +437,7 @@ class TarInputStream
     public void longSkip(final long numToSkip) throws IOException {
       for(long skipped = 0; skipped < numToSkip;) {
         if(numToSkip - skipped > Integer.MAX_VALUE) {
-          skip((int)Integer.MAX_VALUE);
+          skip(Integer.MAX_VALUE);
           skipped += Integer.MAX_VALUE;
         } else {
           skip((int)(numToSkip - skipped));
